move Block into serializers file and update, change Transcript to general Toggle component
